+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.8.16
 
 * added __slots__ to `StrEqHashableMixin`
+* added `safe_sleep` to `TerminableThread`